+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - ripple will unrip if unclick or touchend is outside of list-item
 - `mwc-list`
   - fixed regression in list that broke mwc-select in IE or shady dom.
+- Tabs no longer focus on initialization
 
 ## [v0.15.0] - 2020-05-05